Reorganize retains annotation classes
Since they are just backup syntax, no need to put them in the scala package.
