+++ compiler/src/dotty/tools/dotc/core/Definitions.scala
@@ -1017,8 +1017,8 @@ class Definitions {
   @tu lazy val VarargsAnnot: ClassSymbol = requiredClass("scala.annotation.varargs")
   @tu lazy val SinceAnnot: ClassSymbol = requiredClass("scala.annotation.since")
   @tu lazy val RequiresCapabilityAnnot: ClassSymbol = requiredClass("scala.annotation.internal.requiresCapability")
-  @tu lazy val RetainsAnnot: ClassSymbol = requiredClass("scala.retains")
-  @tu lazy val RetainsByNameAnnot: ClassSymbol = requiredClass("scala.retainsByName")
+  @tu lazy val RetainsAnnot: ClassSymbol = requiredClass("scala.annotation.retains")
+  @tu lazy val RetainsByNameAnnot: ClassSymbol = requiredClass("scala.annotation.retainsByName")
 
   @tu lazy val JavaRepeatableAnnot: ClassSymbol = requiredClass("java.lang.annotation.Repeatable")
