--- library/src-bootstrapped/scala/Retains.scala
@@ -1,7 +1,0 @@
-package scala
-import annotation.experimental
-
-/** An annotation that indicates capture
- */
-@experimental class retains(xs: Any*) extends annotation.StaticAnnotation
-
