+++ tests/neg-custom-args/captures/capt-depfun.scala
@@ -1,3 +1,4 @@
+import annotation.retains
 class C
 type Cap = C @retains(*)
