+++ tests/neg-custom-args/captures/capt-test.scala
@@ -1,3 +1,4 @@
+import annotation.retains
 import language.experimental.erasedDefinitions
 
 class CT[E <: Exception]
