+++ tests/neg-custom-args/captures/capt1.check
@@ -1,51 +1,51 @@
--- [E007] Type Mismatch Error: tests/neg-custom-args/captures/capt1.scala:3:2 ------------------------------------------
-3 |  () => if x == null then y else y  // error
+-- [E007] Type Mismatch Error: tests/neg-custom-args/captures/capt1.scala:4:2 ------------------------------------------
+4 |  () => if x == null then y else y  // error
   |  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   |  Found:    {x} () -> ? C
   |  Required: () -> C
   |
   | longer explanation available when compiling with `-explain`
--- [E007] Type Mismatch Error: tests/neg-custom-args/captures/capt1.scala:6:2 ------------------------------------------
-6 |  () => if x == null then y else y  // error
+-- [E007] Type Mismatch Error: tests/neg-custom-args/captures/capt1.scala:7:2 ------------------------------------------
+7 |  () => if x == null then y else y  // error
   |  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   |  Found:    {x} () -> ? C
   |  Required: Matchable
   |
   | longer explanation available when compiling with `-explain`
--- [E007] Type Mismatch Error: tests/neg-custom-args/captures/capt1.scala:13:2 -----------------------------------------
-13 |  def f(y: Int) = if x == null then y else y  // error
+-- [E007] Type Mismatch Error: tests/neg-custom-args/captures/capt1.scala:14:2 -----------------------------------------
+14 |  def f(y: Int) = if x == null then y else y  // error
    |  ^
    |  Found:    {x} Int -> Int
    |  Required: Matchable
-14 |  f
+15 |  f
    |
    | longer explanation available when compiling with `-explain`
--- [E007] Type Mismatch Error: tests/neg-custom-args/captures/capt1.scala:20:2 -----------------------------------------
-20 |  class F(y: Int) extends A:   // error
+-- [E007] Type Mismatch Error: tests/neg-custom-args/captures/capt1.scala:21:2 -----------------------------------------
+21 |  class F(y: Int) extends A:   // error
    |  ^
    |  Found:    {x} A
    |  Required: A
-21 |    def m() = if x == null then y else y
-22 |  F(22)
+22 |    def m() = if x == null then y else y
+23 |  F(22)
    |
    | longer explanation available when compiling with `-explain`
--- [E007] Type Mismatch Error: tests/neg-custom-args/captures/capt1.scala:25:2 -----------------------------------------
-25 |  new A: // error
+-- [E007] Type Mismatch Error: tests/neg-custom-args/captures/capt1.scala:26:2 -----------------------------------------
+26 |  new A: // error
    |  ^
    |  Found:    {x} A
    |  Required: A
-26 |    def m() = if x == null then y else y
+27 |    def m() = if x == null then y else y
    |
    | longer explanation available when compiling with `-explain`
--- [E007] Type Mismatch Error: tests/neg-custom-args/captures/capt1.scala:31:24 ----------------------------------------
-31 |  val z2 = h[() -> Cap](() => x) // error
+-- [E007] Type Mismatch Error: tests/neg-custom-args/captures/capt1.scala:32:24 ----------------------------------------
+32 |  val z2 = h[() -> Cap](() => x) // error
    |                        ^^^^^^^
    |                        Found:    {x} () -> {*} C
    |                        Required: () -> box {*} C
    |
    | longer explanation available when compiling with `-explain`
--- [E007] Type Mismatch Error: tests/neg-custom-args/captures/capt1.scala:32:5 -----------------------------------------
-32 |    (() => C())  // error
+-- [E007] Type Mismatch Error: tests/neg-custom-args/captures/capt1.scala:33:5 -----------------------------------------
+33 |    (() => C())  // error
    |     ^^^^^^^^^
    |     Found:    ? () -> {*} C
    |     Required: () -> box {*} C
