+++ tests/neg-custom-args/captures/cc1.scala
@@ -1,3 +1,4 @@
+import annotation.retains
 object Test:
 
   def f[A <: Matchable @retains(*)](x: A): Matchable = x // error
