+++ tests/neg-custom-args/captures/try.check
@@ -1,42 +1,42 @@
--- [E007] Type Mismatch Error: tests/neg-custom-args/captures/try.scala:22:49 ------------------------------------------
-22 |  val a = handle[Exception, CanThrow[Exception]] { // error
+-- [E007] Type Mismatch Error: tests/neg-custom-args/captures/try.scala:23:49 ------------------------------------------
+23 |  val a = handle[Exception, CanThrow[Exception]] { // error
    |                                                 ^
    |                              Found:    ? ({*} CT[Exception]) -> {*} CT[? >: box ? Exception <: box ? Exception]
    |                              Required: CanThrow[Exception] => box {*} CT[Exception]
-23 |    (x: CanThrow[Exception]) => x
-24 |  }{
+24 |    (x: CanThrow[Exception]) => x
+25 |  }{
    |
    | longer explanation available when compiling with `-explain`
--- [E007] Type Mismatch Error: tests/neg-custom-args/captures/try.scala:28:43 ------------------------------------------
-28 |  val b = handle[Exception, () -> Nothing] {  // error
+-- [E007] Type Mismatch Error: tests/neg-custom-args/captures/try.scala:29:43 ------------------------------------------
+29 |  val b = handle[Exception, () -> Nothing] {  // error
    |                                           ^
    |                                           Found:    ? (x: {*} CT[Exception]) -> {x} () -> ? Nothing
    |                                           Required: CanThrow[Exception] => () -> Nothing
-29 |    (x: CanThrow[Exception]) => () => raise(new Exception)(using x)
-30 |  } {
+30 |    (x: CanThrow[Exception]) => () => raise(new Exception)(using x)
+31 |  } {
    |
    | longer explanation available when compiling with `-explain`
--- Error: tests/neg-custom-args/captures/try.scala:39:4 ----------------------------------------------------------------
-34 |  val xx = handle {
-35 |    (x: CanThrow[Exception]) =>
-36 |      () =>
-37 |        raise(new Exception)(using x)
-38 |        22
-39 |  } {  // error
+-- Error: tests/neg-custom-args/captures/try.scala:40:4 ----------------------------------------------------------------
+35 |  val xx = handle {
+36 |    (x: CanThrow[Exception]) =>
+37 |      () =>
+38 |        raise(new Exception)(using x)
+39 |        22
+40 |  } {  // error
    |           ^
    |           The expression's type box {*} () -> Int is not allowed to capture the root capability `*`.
    |           This usually means that a capability persists longer than its allowed lifetime.
-40 |    (ex: Exception) => () => 22
-41 |  }
--- Error: tests/neg-custom-args/captures/try.scala:51:2 ----------------------------------------------------------------
-46 |val global = handle {
-47 |  (x: CanThrow[Exception]) =>
-48 |    () =>
-49 |      raise(new Exception)(using x)
-50 |      22
-51 |} {  // error
+41 |    (ex: Exception) => () => 22
+42 |  }
+-- Error: tests/neg-custom-args/captures/try.scala:52:2 ----------------------------------------------------------------
+47 |val global = handle {
+48 |  (x: CanThrow[Exception]) =>
+49 |    () =>
+50 |      raise(new Exception)(using x)
+51 |      22
+52 |} {  // error
    |             ^
    |             The expression's type box {*} () -> Int is not allowed to capture the root capability `*`.
    |             This usually means that a capability persists longer than its allowed lifetime.
-52 |  (ex: Exception) => () => 22
-53 |}
+53 |  (ex: Exception) => () => 22
+54 |}
