+++ tests/pos-custom-args/captures/byname.scala
@@ -1,3 +1,4 @@
+import annotation.retainsByName
 class CC
 type Cap = {*} CC
 
@@ -7,4 +8,6 @@ def test(cap1: Cap, cap2: Cap): {cap1} I =
   def f() = if cap1 == cap1 then I() else I()
   def h(x: {cap1}-> I) = x
   h(f()) // OK
+  def hh(x: -> I @retainsByName(cap1)) = x
+  h(f())
