+++ tests/pos-custom-args/captures/capt0.scala
@@ -2,6 +2,6 @@ object Test:
 
   def test() =
     val x: {*} Any = "abc"
-    val y: Object @scala.retains(x) = ???
-    val z: Object @scala.retains(x, *) = y: Object @scala.retains(x)
+    val y: Object @scala.annotation.retains(x) = ???
+    val z: Object @scala.annotation.retains(x, *) = y: Object @annotation.retains(x)
