+++ tests/pos-custom-args/captures/list-encoding.scala
@@ -1,4 +1,5 @@
 package listEncoding
+import annotation.retains
 
 class Cap
